requirements.txt
# Core data science
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
imbalanced-learn>=0.11.0
xgboost>=2.0.0

# Visualization (notebook only)
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
pyyaml>=6.0
joblib>=1.3.0

# Optional: for Milestone 4 deployment
# fastapi>=0.110.0
# uvicorn>=0.27.0
# pydantic>=2.0.0
